Packaging updates for 2022.1.0

+++ buildWinDistributions.ps1
@@ -1,5 +1,6 @@
 #! powershell
 
+param ($install_pp=1)
 # build simple distributions
 # python setup.py bdist_egg
 # python setup.py sdist --formats=zip
@@ -21,9 +22,11 @@ for ($i=0; $i -lt $pyPaths.Length; $i++) {
     [console]::beep(440,300); [console]::beep(880,300)
     # try to uninstall psychopy from site-packages
     # re-install the current version as editable/developer
-    Invoke-Expression ("& '{0}python.exe' -m pip install . --no-deps --force" -f $pyPaths[$i])
-    echo ("Installed current PsychoPy")
-    xcopy /I /Y psychopy\*.txt $pyPaths[$i]
+    if ($install_pp -eq 1) {
+        Invoke-Expression ("& '{0}python.exe' -m pip install . --no-deps --force" -f $pyPaths[$i])
+        echo ("Installed current PsychoPy")
+        xcopy /I /Y psychopy\*.txt $pyPaths[$i]
+    }
 
     # build the installer
     $thisPath = $pyPaths[$i]
@@ -34,12 +37,13 @@ for ($i=0; $i -lt $pyPaths.Length; $i++) {
     Invoke-Expression $cmdStr
     # "C:\Program Files\Caphyon\Advanced Installer 13.1\bin\x86\AdvancedInstaller.com" /rebuild PsychoPy_AdvancedInstallerProj.aip
 
-    Invoke-Expression ("& '{0}python.exe' setup.py clean --all" -f $pyPaths[$i])  # clean up our build dir
-    # try to uninstall psychopy from site-packages
-    Invoke-Expression ("& '{0}python.exe' -m pip uninstall -y psychopy" -f $pyPaths[$i])
-    # re-install the current version as editable/developer
-    Invoke-Expression ("& '{0}python.exe' -m pip install -e . --no-deps" -f $pyPaths[$i])
-
+    if ($install_pp -eq 1) {
+        Invoke-Expression ("& '{0}python.exe' setup.py clean --all" -f $pyPaths[$i])  # clean up our build dir
+        # try to uninstall psychopy from site-packages
+        Invoke-Expression ("& '{0}python.exe' -m pip uninstall -y psychopy" -f $pyPaths[$i])
+        # re-install the current version as editable/developer
+        Invoke-Expression ("& '{0}python.exe' -m pip install -e . --no-deps" -f $pyPaths[$i])
+    }
 }
 
 echo 'moving files to ..\dist'
